pull exception into separate file
