+++ src/sentry/grouping/fingerprinting/exceptions.py
@@ -1,0 +1,2 @@
+class InvalidFingerprintingConfig(Exception):
+    pass
